experimental search input: Blur input on submission (#49048)
**NOTE:** Depends on #48924. Review with whitespace changes hidden.

Currently the search input stays focused when the query is submitted,
which implies showing suggestions. This is a problem on the search
results page because it covers up part of the page. While the input will
be blurred automatically as soon as the results arrived, the input
itself shouldn't assume that it will be blurred via an external event.

This PR updates the code to blur the input when the query is submitted.
It also fixes a bug where the input is refocused (and therefore the
suggestions are shown again) after the query is submitted when the
cursor is not positioned at the end of the input.
I wasn't able to find out _why_ this happens but I know it's related to
the fact that we update the selection in
`useUpdateEditorFromQueryState` (if I comment out the `selection` change
in line 416 the input is not focused again).
To avoid this behavior I simply avoid changing the selection when the
query hasn't changed and when the selection is not explicitly set.


Before/after video: 


https://user-images.githubusercontent.com/179026/224158472-fefe36a3-99cf-4a6d-99ff-50beab091468.mp4



## Test plan

Go to search results via an arbitrary search. Position the cursor within
the query (not at the end!). Press enter to submit the query
  -> The query input blurs and stays blurred.

## App preview:

-
[Web](https://sg-web-fkling-search-input-blur-on-submit.onrender.com/search)

Check out the [client app preview
documentation](https://docs.sourcegraph.com/dev/how-to/client_pr_previews)
to learn more.