experimental search input: Fix cursor position in Chrome
Chrome moves focus to the input even when clicking on the container
element around it. Clicking above the input will put the cursor at the
beginning of the input.
This isn't useful because with the context: filter being present by
default because the user most likely wants to position the cursor after
the context: filter to type the query.

It's reasonable to assume that it's not uncommon for users to click
slightly outside the input.

Debugging revealed that we can identify such transactions (not 100%
uniquely but close enough). This PR adds a transaction filter to adjust
the cursor position to the end of the input in this case.

See code for more comments.