experimental search input: Show example query/filter suggestions (#49115)
This PR adds a new suggestion source: examples. If the query is empty or
the user is at a whitespace token we are showing up to three examples at
the bottom to expose users to more advanced features.

When the user selects one of these examples (or explicitly hides it) we
mark it as "used" and don't show it again.

The examples are randomized every 10 minutes or whenever the list needs
to be updated to contain 3 items.



https://user-images.githubusercontent.com/179026/224369695-5839bc66-26fc-45a9-a5cd-13895f2b376d.mp4


This PR also refactors the default and filter suggestions and adds
`select:` and `OR` to the default suggestions.


## Test plan

- Focus query input. Suggestions should show examples by default.
- Select an example -> input gets updated
- "Hide" an example -> example is removed and suggestions list gets
updated

I added `console.log` statements to verify that the used examples are
stored correctly in temporary settings.

## App preview:

-
[Web](https://sg-web-fkling-search-input-onboarding.onrender.com/search)

Check out the [client app preview
documentation](https://docs.sourcegraph.com/dev/how-to/client_pr_previews)
to learn more.