experimental search input: Fix cursor position in Chrome (#49154)
Addresses [this slack
thread](https://sourcegraph.slack.com/archives/C03CSAER9LK/p1678463340977319)

Chrome moves focus to the input even when clicking on the container
element around it. Clicking above the input will put the cursor at the
beginning of the input.
This isn't useful because with the context: filter being present by
default because the user most likely wants to position the cursor after
the context: filter to type the query.

It's reasonable to assume that it's not uncommon for users to click
slightly outside the input.

Debugging revealed that we can identify such transactions (not 100%
uniquely but close enough). This PR adds a transaction filter to adjust
the cursor position to the end of the input in this case.

See code for more comments.

Before/after video: 


https://user-images.githubusercontent.com/179026/224445816-72eaacfd-bd1a-4ee3-a5d8-aba410655436.mp4



## Test plan

- Click slightly outside of input in Chrome -> input gets focus with
cursor being positioned at the end of the input.
- Click on text inside the input -> cursor is positioned where the click
happened
- Move cursor with arrow keys -> cursor moves as expected

## App preview:

-
[Web](https://sg-web-fkling-search-input-smart-focus.onrender.com/search)

Check out the [client app preview
documentation](https://docs.sourcegraph.com/dev/how-to/client_pr_previews)
to learn more.