app: implement notifications sent by sourcegraph.com to Sourcegraph App users (#49143)
This gives us the ability to add notifications for Sourcegraph App users
to the Sourcegraph.com site configuration (under
`dotcom.app.notifications`) with Markdown messages including text, links
& more.

When the App sends a telemetry ping, the response now includes these
notifications, which the app then stores in the site config
`notifications` field. They are then rendered as dismissible alerts:

<img width="714" alt="image"
src="https://user-images.githubusercontent.com/3173176/224429630-f00d012b-118d-4dab-9a08-9ad2c54d237f.png">

Each notification has a unique key, and so once dismissed the user will
never see it again (unless we add a new one in the sourcegraph.com site
config.)

Additionally, we can specify a version range in the sourcegraph.com site
config to make the notification target specific Sourcegraph App
versions. For example, if we release a broken version we could notify
those users specifically and inform them.

## Test plan

Because this hinges on a sourcegraph.com change, this is hard to test.
What I've done to de-risk it is:

1. Made each commit individually clear/transparent; most of the changes
to the `updatecheck` code are VERY low risk and easily verified.
2. Every change only affects App; there is near-zero risk to other
deployment types.
3. Once merged; we can test this further (post-code-freeze) and confirm
it works. We still have time to land bug fixes should any issue pop up.
4. I used VERY defensive coding; so if it doesn't work the notifications
will just not appear (rather than e.g. crash the app, or cause pings to
stop working.)

I think this is worth the risk in terms of value it gives us, and
because it makes pings more defensible to users, and gives us a better
way to communicate with these early adopters. I suggest we land this
before code freeze as a result - even if it's not 100% perfect.

---------

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>