(experimental) search input: Explicitly blur input on repo pages
Currently when navigating to a repository page the search input
continues to have focus. That's not a big problem in the current search
input because it doesn't take up much space. The new input however will
always show suggestions when focused, covering parts of the page.

This commit makes two changes:

- Blur the input when a "goto" action is executed
- Add a blur editor hint used by the repository page container to
  instruct the editor to blur

The latter is necessary because for unknown reasons the input regains
focus when the repository page updates the input to reflect the
currently opened repo/file.