SCIM: Welcome email when new scim user created (#49052)
Adds a new welcome email that is based on the invitation email but does
not include who invited you since the action was done automatically when
the user was given permission to Sourcegraph from SCIM (via an IDP such
as Okta).

The welcome email re-uses the env variable that is used to disable
invite emails. This felt like the logical choice, thinking the user
would expect the "welcome" and "invite" emails to act in the same
manner.

**note** - Using tools like the SCIM validator will attempt to send live
emails, be sure to either set `DISABLE_EMAIL_INVITES` or
`DEBUG_EMAIL_INVITES_MOCK` when running live SCIM api calls and you do
not want real emails to go out.