experimental search input: Remove goto support for file and symbol suggestions
File/symbol _completions_ don't work well together with file/symbol
_go to_ support:

- to make _go to_ useful we need to show an option for every unique
  file/symbol
- this means we could have multiple files/symbols with the same name
  (from different repos/files)
- selecting any of these duplicate options for _completion_ fill have
  the same affect on the query input (e.g. adding a `file:` filter)

Having multiple options perform the same action is possibly confusing to
the user because might not be obvious to them why they all behave the
same (or it might not be obvious which one to choose).

So for now this removes the option to go to files and symbols from
suggestions, and removes the repo/file descriptions until we find a
clearer way to make this functionality available.