symbols: use shared pool of connections across all gRPC methods (#49041)
This adds gRPC connection caching to symbols. To do so, it introduces a new generic TTL cache. We implement connection caching this way so that we don't need to dig too deeply into the conf package, which is messy. This allows us to make an implementation that is entirely decoupled from conf.