experimental search input: Remove goto support for file and symbol suggestions (#49218)
File/symbol _completions_ don't work well together with file/symbol _go
to_ support:

- to make _go to_ useful we need to show an option for every unique
file/symbol
- this means we could have multiple files/symbols with the same name
(from different repos/files)
- selecting any of these duplicate options for _completion_ will have
the same affect on the query input (e.g. adding a `file:` filter)

Having multiple options perform the same action is possibly confusing to
the user because might not be obvious to them why they all behave the
same (or it might not be obvious which one to choose).

So for now this removes the go to feature for files and symbols, removes
the repo/file descriptions and deduplicates the file and symbol options
(until we find a clearer way to make this functionality available).



![2023-03-13_14-29](https://user-images.githubusercontent.com/179026/224763268-eebc9646-5de1-44af-9343-8127041a9298.png)



## Test plan

Type into the query input. File and symbol suggestions don't have a
"goto" action or description anymore. The footer does not information
about `ctrl/cmd+enter` anymore.