SCIM: Add Azure specific logic for PATCH endpoint (#49200)
Azure AD deviates from the SCIM spec for PATCH requests against
multi-value attributes with a filter. Azure expects "replace" operations
to act as "add" operations if no item matches the filter. The spec
states:
```
If the target location is a multi-valued attribute for which a value selection filter ("valuePath")
has been supplied and no record match was made, the service provider SHALL indicate failure
by returning HTTP status code 400 and a "scimType" error code of "noTarget".
```

This change adds a new site setting to select the Identity provide and
logic to the PATCH endpoint to support this behavior only when Azure AD
is set as the current Identity Provider.

Co-authored-by: David Veszelovszki <veszelovszki@gmail.com>