build-tracker: keep better track of jobs (#48378)
* Build has moved to its own package, and instead of keeping track
individual jobs, the Build now keeps track of Steps.
* Steps accumulate Jobs with the same name. A Step is able to determine
its `FinalStatus`, which is done by looking at the last two Jobs to
determine whether the step has Passed, Failed or is Fixed.
* All things Notification like has been moved to the notify package. The
notifiy client no longer keeps track of teammates, instead, it keeps
track of notifications that have been sent for a build. ~If a build
wants to send an updated notification, the teammate is reused from the
previous build.~ In a follow up PR I will address this more thoroughly.
* The notification client now only works with a `BuildNotification` and
has no access to the original build. This makes the boundary between
keeping track of a build and sending a notification very clear. The only
way to send a notification is to provide all the information in a
BuildNotification

## Test plan
Unit Tests and Integrations tests see this channel
https://sourcegraph.slack.com/archives/C03MM9C63KQ
<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->