authz/azuredevops: List user's orgs before listing repos
This is an optimization for the number of API calls we make for each
user permissions sync in Azure DevOps, which will help customers with a
long list of orgs and users.

Instead of attempting to list repos of every org, we will list the orgs
that the user has access to and then iterate over those.