dbtest: add isolation check for test databases
Because we're running stateful agents, some databases might not be
deleted afterward. On a normal scenario, this won't happen, but the agent
container might crash, or someone is running an old commit, and it might
still happen. Therefore, we're being defensive here to ensure it doesn't
happen, because it would lead to a CI failure.

[force-bazel]