App: Fix UI flickering of limited access app banners
Reported by Vova in https://github.com/sourcegraph/sourcegraph/pull/49220#pullrequestreview-1337564414

This PR fixes an issue where using `useTemporarySettings()` causes the
UI to flicker and render the banner for a frame until the observable is
properly initialized.

We need to fix this properly in the future (maybe by migrating the hook
to use Apollo and the new suspended query). For now we store if banners
are closed in local storage instead.