Ranking: add monitoring dashboards for ranking quality (#49275)
This PR updates the monitoring dashboards related to search ranking:
* Replaces the "mean click position" dashboard with two that are more
robust. The mean position can be noisy and swayed by outliers. Now, we
report the percentage of clicks on the top result, and the percentage
within the top 3 (roughly the first page). These represent ranking
successes. We exclude searches with very few results, since a click on a
top result is not meaningful when there are no other results.
* Corrects an existing dashboard about the distribution of search clicks
by result type.
* Adds a new dashboard with the percentage of Zoekt searches that hit
the flush time limit. In our experiments on dot com, we found this
really helpful in understanding how many searches could be missing
relevant results or non-deterministic.

## Test plan

Ran `sg start monitoring` locally, ran example searches, and inspected
dashboards.