[chore] disallow adding authorization on dotcom
This changes previous behavior, where authorization was enforced
on dotcom for github and gitlab code hosts. We do not have any
private code on dotcom, show error when you try adding a private
repo to dotcom and this is the final nail in the coffin, by not
even allowing to specify that you want to enforce permissions
on dotcom.