gitserver: pass coursier cache dir down to jvmpackages (#49453)
Before this change we would `env.Get` register the same `SRC_REPOS_DIR`
in two locations (and have to keep the descriptions and default values
in sync), which is a bit nasty. For App, it also caused two problems:

1. Init order issue - one location was done as part of the service
config loading, while the other was in a global `var()` block. The
latter doesn't give App a chance to set the default value.
2. `panic: env var "SRC_REPOS_DIR" already registered with a different
description or value` (due to the default values differing due to point
2 above)

This change makes us pass the coursier cache dir down to jvmpackages
from the gitserver svc config, which I think is nicer code quality &
fixes the app issues above.

## Test plan

existing tests + careful review

---------

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>