move nonce into self contained struct
I think this simplifies reasoning around the state of nonce and makes it
clearer the code is safe / correct.

This has one possible downside: if our login logic after succesfully
verifying the nonce fails the user won't be able to retry. This seems
rare.

One upside is before we technically had replay attacks. But in practice
for app this is not an issue.