Smartsearch: recognize alternates in capture groups as regex (#49486)
With the current smartsearch rules, we don't optimistically execute
patterns like `(a|b)` because our heuristic doesn't recognize it as
regexy enough to be confident it's a regex pattern. However, an
alternation inside a capture group would be a pretty rare pattern to see
outside an attempted regex search, so let's add slightly special case to
treat that pattern as regex.