insights: chore: link to pings list in docs from site admin pings page (#49533)
I noticed that 1) the list of Code Insights pings in the site admin page
was not up to date at all and 2) it's just essentially a reformat of the
list of what's in the docs. Ideally we could just pull that information
directly.

I don't want to make the change for all other types of usage data but I
think it makes sense to just link to the docs. We shouldn't have to
maintain yet another list of pings we have.

## Test plan

Visual check