git-combine: create synthetic origin repo in tmpdir (#49536)
As part of bazel migration we stopped using the checked out sourcegraph
repo and introduced this synthetic origin repo. This commit makes it so
we only use the synthetic repo if we don't have access to a local
repository.

Test Plan: Ran inside of the sg repo and outside of a git repo. In the
sg repo you can see in the output it doesn't create a repo and it takes
longer for it to run git fetch (since it fetches 100 real commits
instead of 2 synthetic ones).

``` shell
cd internal/cmd/git-combine
go test -c .
./git-combine.test -test.v

mv git-combine.test /tmp
cd /tmp
./git-combine.test -test.v
```