insights: during backfill skip not found repos (#49548)
During the time between the repos being selected for an insight series
and attempting to run the backfill queries, it is possible that repos
are marked as deleted. Currently when this happens the error is captured
are recorded against the insight. This causes a UI element to appear on
the insight warning the user that the insight does not have all the
data.

When the insight is not found we can continue processing without
recording an error because a search against this repo would return no
results, so there is not any missing data.