SCIM: Fix some random issues (#49393)
- **Giving 404 response for invalid IDs**: Okta validator sends stuff
like `010101001010101011001010101011` to check nonexistent users. We
returned `400` to these, which made the tests fail. We now return `404`.
- **Simplified docs**: The `scim.identityProvider` setting was mentioned
as required, and its description wasn't that specific and led me to
believe that this setting was more than just "standard/Azure AD". I've
marked it as optional, and cleared up its description.
- **Renamed two constants** to match Go standards. (My IDE gave a
warning on these.)
- **Deleted unused arguments** to get rid of code warnings.
- I also **fixed some typos** in comments.
- Tested with Azure AD validator and Okta's SPEC validator.