[fix] properly handle public repos with explicit perms and unified model (#49576)
### Previous behavior

[As seen
here](https://sourcegraph.sourcegraph.com/github.com/sourcegraph/sourcegraph@1fc50a2878c17dd0199a44acbe521726c425b449/-/blob/internal/database/repos_perm.go?L164-167)
when explicit permissions are **ON**, we are turning **OFF** the
condition that checks if the external service is restricted. In the same
conditional logic, we are also turning **OFF** the check if the repo is
public.
Which basically means, that if explicit permissions are ON, we are
treating public repos as private.

That does not sound right to me 😞

### With the fix

We use the condition for public repos and unrestricted external service
even when explicit permissions are enabled. But _ONLY_ if also the new
unified perms model is enabled.

This way we limit the exposure to instances that enabled the
experimental feature `unifiedPermissions`.

## Test plan

Tested locally + unit tests
The feature is behind a feature flag, so expect no impact on customers.

(cherry picked from commit 8d57029e50193428cbe38ef084ea7e1008450f7e)