Code navigation: impose minimum latency on goto definition
Previously, there was no visual hint in the UI when the user cmd-clicked
multiple times from the same location. This behavior felt buggy because
it looks like the app is unresponsive (nothing happens) when in reality
the action happened too fast and nothing changed because you already
triggered the action. This situation happens frequently when the
definition is already visible in the viewport and the user may not have
noticed where the focus moved.

In this PR, we delay "Go to definition" for maximum 20 milliseconds when
there is only a single definition. There is no delay if the request
takes more than 20 milliseconds to resolve.  This 20 millisecond delay
makes the focus on the definition token disappear briefly before it
re-appears again making it easier to notice that something happened on
cmd-click.