Don't access process.env in browser build
Different from webpack, esbuild does not patch the global object and
create a `process` object with all environment flags. Instead, they will
be replaced at compile time (which I think is actually more sane because
we are not in a node process environment anyways).

Because of this though, we can't simply access `process.env` because it
might be undefined. Instead of changing the esbuild output, this PR adds
a guard to check of `process` is defined on the global object. This is
always the case in a Node env.