insights: Update fetch policy for all insights view so it picks up new insights (#49612)
closes https://github.com/sourcegraph/sourcegraph/issues/44960

When creating a new insight they are not visible on the All Insights
page until the user refreshes the page, because the page is relying on
the cache. This change instructs the page to return with the cache but
also run a new network request to update data.

(cherry picked from commit af72969e4974a3daf4e3890b7d51ec6bc77f98ee)