No team cycles (#49556)
This change introduces a check on `UpdateTeam` to not introduce a
circular parent relationship.

The check effectively finds all teams that _are not descendants_ (so not
in the subtree of) modified team, and looks whether the parent is there.
It is only safe to set the parent team ID to a team outside of modified
team's own descendants.

## Test plan

*   Testing database stores with Postgres.
*   Testing GraphQL API with fakes.

(cherry picked from commit da923735d2f61edb71545e200116c34910e9f7bd)