Highlighting: use same color for built-in identifiers and keywords
Previously, we used different colors for built-in identifiers and for
keywords. On top of that, we didn't consistently use different kinds of
colors between dark and light themes.

In this PR, we change the colors for all built-in identifiers to have
the same color as keywords. This color choice is consistent with how
JetBrains IDEs color tokens.