executor: Add shell runtime and fat image (#49460)
This image is not meant for production use. It enables the use of some
very basic batch changes when using the native-SSBC feature flag, but is
not isolated and thus should not be used.

More context on why we're doing this:
https://sourcegraph.slack.com/archives/C07KZF47K/p1678954730651559?thread_ts=1678480841.394689&cid=C07KZF47K

We should undo this change once we're in a better situation for the
reason of this PR.

(cherry picked from commit d516e8a7b6e7df631a74cbda0fda384eeff28d5a)