experimental search input: Scroll selection into view after completion
Originally I wanted to do properly pass in the `globbing` parameter but
then I noticed that this feature had been removed for quite a while
(#27886, #46045).
But while working on this I noticed that the cursor is not properly
scrolled into view when completing a long value. Adding `scrollIntoView`
fixes that.