App Prototype: Add "Connect to Sourcegraph.com" prototype (#49560)
Part of #47998

This uses #49107 to add a new "Connect to Sourcegraph.com" workflow to
App. When no access token is set in the site config, an option to
"Connect to Sourcegraph.com" is rendered in the menu dropdown. When
clicked, the user is redirected to dotcom where an auth token is created
and the user is directed back to App.

To check if a user is signed-in, I added a new GraphQL endpoint so that
we can, in the future, make sure the token is valid by sending a POST to
`https://sourcegraph.com/.api/graphql` to validate the token.

The redirect screen handles invalid `code` parameters in the URL and
uses a full page reload after the token is written to the site config to
avoid any stale cache issues (since this is localhost only, it's very
fast anyways).

## Test plan


https://user-images.githubusercontent.com/458591/225983059-4e2830b3-a332-4ec1-b9e4-88e376330561.mov

(My local devserver runs on a different port than App so there's an
awkward manual URL replacement happening)

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->