[fix] do not set providerState on FetchAccount success
This pollutes our logs with needless information, we do not care that
FetchAccount succeeded, we only care about errors. Most of the provider
implementations return nil from this function anyway.

It also pollutes the result of the sync job, since there was a success
message for the FetchAccount. We recently added logic in #49752
to treat all errors for provider states as an error and this
success message broke that logic.