cody: refactor architecture (#49707)
This is my proposal for a new Cody extension architecture. It might not
be the ideal architecture, but it's a huge improvement over the current
situation.

Here are the important concepts:

* `Interaction` is the atomic unit of the chat. It contains the human
question, the assistant response, and the context, if necessary. An
interaction can be converted to a prompt or a (chat) UI message.
* `Transcript` is a linear collection of `Interaction`s that can be
converted to a prompt or a full UI conversation. It also handles
limiting the prompt to a maximum length.
* **Everything is a recipe**, including the chat questions. A recipe
receives the necessary dependencies (chat input, editor, intent
detection, and codebase context) and constructs an `Interaction`. The
`ChatViewProvider` handles executing the recipe and adding the resulting
`Interaction` to the `Transcript`.
* Context fetching has been abstracted into the `CodebaseContext` class.
It handles switching between `embeddings` and `keyword` searching.
`blended` context type has not been implemented yet.
* Context messages are included for the last message only. See unit
tests for examples.
* Prompt structure is now enforced through unit tests.

## Test plan

* Chatting and executing recipes should still work.
* `pnpm run build && pnpm run test:integration && pnpm run test:unit`
should be green.

## App preview:

- [Web](https://sg-web-rn-cody-refactor-extension.onrender.com/search)

Check out the [client app preview
documentation](https://docs.sourcegraph.com/dev/how-to/client_pr_previews)
to learn more.