users: fix nil panic
The `LogPasswordEvent` function can be called with `nil` as the request.
This fixes the nil panic from this call: https://sourcegraph.com/github.com/sourcegraph/sourcegraph/-/blob/internal/database/users.go?L1531:3

It also removes the deafult values that don't need to be set to remove
references to the field.