Build repo-updater with Bazel instead (#49228)
This PR patches to `build.sh` for `repo-updater` so it builds with Bazel
with `$DOCKER_BAZEL==true`.

This is the first step of speeding up the builds by moving the docker
builds to use Bazel. While there are better ways of doing this,
timewise, this is the fastest approach.

Once we have this in place and have rolled out Bazel, we'll be able to
investigate using the `rules_docker` approach to take this further.

This can serve as an example for porting other Dockerfiles and we can
then introduce a single job that does all of this. Whether or not we
parallelize it is to be studied once we have more insights on how quick
it builds on a stateful agent.

## Test plan

please note that `VERSION` must be exported, not just set. 

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

<img width="1009" alt="CleanShot 2023-03-13 at 18 22 59@2x"
src="https://user-images.githubusercontent.com/10151/224778509-e5cb3c60-8994-4c2d-a2e0-a823aaa649eb.png">

![CleanShot 2023-03-13 at 18 23
08@2x](https://user-images.githubusercontent.com/10151/224778766-e0678df9-ef7e-4af8-9671-4321f32dae42.png)

Co-authored-by: davejrt <davetry@gmail.com>