Add index to speed up deletions from codeintel_ranking_references
There's a FK to this table on the codeintel_ranking_references_processed table but no suitable index to check this constraint so it's effectively doing a seq scan over that table for every deleted record. This fixes it. Deleting 100 records went from 10s to 20ms.
A query to delete 8.5M records that was running for 2d now and didn't finish yet finished in 2 or 3 minutes now.