users: fix nil panic (#49779)
The `LogPasswordEvent` function can be called with `nil` as the request.
This fixes the nil panic from this call:
https://sourcegraph.com/github.com/sourcegraph/sourcegraph/-/blob/internal/database/users.go?L1531:3

It also removes the deafult values that don't need to be set to remove
references to the field.

The regression was introduced in
https://github.com/sourcegraph/sourcegraph/pull/49174

## Test plan

- Existing tests and manual testing

(cherry picked from commit 3c0985ca939196e6d7bc215857900507f3aeb967)