all: remove frivolous errors from console for App (#49801)
Prior to this change there were quite a few noisy errors/warnings in the
console when App starts that do not really impact the user. This PR
removes them.

Additionally, there is a small bugfix here where we would initialize the
site admin account and it would 'fail' because the site was already
initialized. In the case of app, this is not an error and was preventing
the browser nonce URL from generating.

Helps #47992


## Test plan

Existing tests + ran `SRC_LOG_LEVEL=warn sg start app` and looked at
what warnings were displayed.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>