permissions-center: remove feature flag and old perms syncing code.
This commit only removes the code without moving it around for simplicity. Moving the code closer to `perms_syncer_worker` and deleting `perms_syncer.go` file will be done in a following commit for the sake of minimizing PR review complexity.

Test plan:
Local sg run and confirming that permissions are being synced.