Fix useObservable
This fixes an issue with `useObservable`. The previous implementation
would always render with the incorrect `observable` during the first
render pass which caused some significant layout flickering all over the
app.

Here's an example where stale data causes the content list to disappear
between navigation of the tree view:

To fix this, we need to make sure the setCurrentValue callback is called
before we yield the rendering to the browser so that the content is in
sync with the observable.