[fix] code host rate limits must be synced first after config change
Previously we synced the rate limits based on the configuration of the
code host only after the code host config is validated, usually via
an HTTP call to the code host API itself.

However if the rate limit is too strict, this call might fail. Applying
the rate limit first fixes situations like
1. Apply too strict rate limit
2. All rate limit is consumed
3. Update the rate limit to higher value
4. the code host config is not applied, because rate limit is exceeded
and the HTTP call to code host fails, hence we never get to applying the
new rate limit

This genuinely happened to us when testing rate limits on rctest, we needed
to restart the repo-updater service completely for the new rate limit
config to be applied.