Fuzzy finder: boost ranking of symbols with shorter names
Previously, the "Symbols" tab of the fuzzy finder didn't use any ranking
to break ties between symbols that had the same fuzzy matching score.
This was problematic because if you wanted to match against a symbol
with short names then you could encounter a situation where you needed
to use the arrow keys to move through the results to reach the result
because you can't narrow down the query the match the symbol with the
short name.

This PR changes the ranking to break fuzzy matching ties by prioritizing
symbols with short names. If the user wants to match against a symbol
with a longer name then the user can add more characters to the query.