site-config: Remove reproducedDiff and make diff non-nullable
In this commit we trip and modify the site config change API:

- With the changes in the backend to skip redundant entries to the
critical_and_site_config table, we no longer require the reporoudcedDiff
boolean to indicate if we were able to reproduce a diff or not. We
simply do not return elements where we cannot reproduce a diff any more.

- Consequently, it is now guaranteed that diff will be non null for every
node in the response.

Finally, we update the frontend compoenent to:
- Not check reproducedDiff
- Not handle null diff