file view: Fix in-file search behavior wrt selections
It's possible to search for word "under the cursor" by selecting the
word/text and then triggering CodeMirror's search. However our custom UI
doesn't integrate well with this behavior: CodeMirror will highlight the
visible matches but our UI doesn't reflect the match state (the input is
empty and no result count is shown). It's possible to go through the
matches with the next/previous buttons though.

The root issue seems to be that we are not deriving the search matches
map and the currently selected match from the searchQuery but only
compute it when we detect changes in the input. After trying out a
couple of things I believe the solution implemented here is more robust.

- Remove state fields for matches and selected index in favor of
  computing those when the searchQuery changes.
- Update the selected index when the selection changes. This also means
  that the index is cleared when the user selects a different text while
  the search is open. Given that CodeMirror changes the highlighting
  though I'd say this is the more correct behavior.