search results cache: fix regression from useObservable changes (#49937)
Fixes a regression in the search result cache that originated as a
regression from https://github.com/sourcegraph/sourcegraph/pull/49838

The search results cache is used primarily in back-navigation scenarios:
you make a search, click a result, and then click the back button; the
same results should be rendered again without executing a new search and
potentially changing the order of the results.

The order each React hook ran in was required to be in a specific order
to make the cache work. The changes to the `useObservable` hook made
that issue surface. This PR fixes that by clarifying the dependencies
and removing one of the `useEffect`s to instead run its side-effect
directly in the observable handling code. It also changes the cache to
be stored in with `useRef` instead of `useState`, as re-renders are not
desired when updating the cache.


## Test plan

Back button navigation:
1. Execute a search
2. Click on a result
3. Click back
Expected: Previous results should load, and no new search should be
executed.

New search after clicking results:
1. Execute a search
2. Click on a result
3. Execute a new search from the search box at the top (don't reload the
page)
Expected: The new search should be executed.

Re-execute the same search:
1. Execute a search
2. Click on the search box and press "Enter"
Expected: The same search should be re-executed and results should be
loaded from scratch.

## App preview:

- [Web](https://sg-web-jp-resultcacheobservablefix.onrender.com/search)

Check out the [client app preview
documentation](https://docs.sourcegraph.com/dev/how-to/client_pr_previews)
to learn more.