make access tokens more identifiable with sgp_ prefix (#49989)
Access tokens now begin with the prefix `sgp_` to make them identifiable
as secrets. You can also prepend `sgp_` to previously generated access
tokens, although they will continue to work as-is without that prefix.
It is also allowed to *omit* the `sgp_` prefix from an access token, for
simplicity of backcompat (so that we do not need to distinguish between
access tokens created prior to the introduction of this prefix).

See
https://github.blog/2021-04-05-behind-githubs-new-authentication-token-formats/
for an explanation of why GitHub made a similar change. Slack, Stripe,
and many other companies have also made this change.

At some point, we should warn, deprecate, and eventually disable access
tokens that do not contain this prefix.



![sgpaccesstoken](https://user-images.githubusercontent.com/1976/227696764-c35bdcc8-41df-47bc-89cf-707e2c5af05a.png)

(Note: This access token was deleted prior to this screenshot being
uploaded.)


## Test plan

Create a new access token. Confirm it contains the `sgp_` prefix.
Confirm it works. Confirm it works without the `sgp_` prefix as well
(for backcompat).