Introduce a bunch of soft failing bazel test target for the main branch.  (#49966)
This introduces a bunch of stable test targets for the main branch. They
are soft-failing, but they will stress test our autoscaling fleet of
bazel agents as well as warming up the cache for us.

## Test plan

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

Will see some builds on main, but even it goes wrong, it won't fail
anything.