gitserver: use -z for ListFiles (#50023)
If you do not pass -z to git ls-tree, it will quote filenames which
contain special characters. This API isn't used much, so we have never
noticed this issue before (most client code uses LsFiles). However, we
Cody is using this API, and that is how we noticed it. Autoindexing uses
it, and likely didn't encounter issues because the pattern passed in was
likely specific enough to not include filenames with special characters.

In the future we should look into unifying ListFiles and LsFiles, but
they are different enough right now. Right now we copy-paste how LsFiles
parses the output.

Test Plan: added a test case