executors: add aws and google executors to release tool (#49651)
Added the following commands to the release tool in order to add the
executor release to the release tool process.
```
'release:prepare-google-executors'
 'release:prepare-aws-executors'
 'release:google-executors-tags'
 'release:aws-executors-tags'
```
The prepare functions create the PR for the changes to update the
version within the files of the repos
 the tag functions publish the tags for the release
 
Still need to implement a verification function to confirm tags have
been published
 

## Test plan
ran command for `prepare-google-executors` in dry run mode. It pushed I
branch looked at the preview of what diff of the PR are would be and all
the changes manually.

`google-executors-tags` command simply runs the ./release.sh command,
which I have run manually before and does publish the tags

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

## App preview:

- [Web](https://sg-web-mucles-release-add-executors.onrender.com/search)

Check out the [client app preview
documentation](https://docs.sourcegraph.com/dev/how-to/client_pr_previews)
to learn more.